- Adding getFullName() for DatabaseObject - Workaround for Oracle driver bug described in http://issues.apache.org/jira/browse/DDLUTILS-29?page=all - Workaround for Oracle driver bug where the column precision is returned as a long, not an int
git-svn-id: http://svn.code.sf.net/p/schemacrawler/code@53 3243f2e7-4e1b-42d2-9742-d984016aa316

+++ schemacrawler/src/main/java/schemacrawler/tools/operation/OperatorCSVOutput.java
@@ -58,7 +58,7 @@ final class OperatorCSVOutput
    *      java.lang.String, java.lang.String, int, java.lang.String)
    */
   public void handleTable(final int ordinalPosition, final String tableName,
-                          final String tableType, final int count,
+                          final String tableType, final long count,
                           final String message)
   {
     out.print(FormatUtils.escapeAndQuoteForExcelCsv(tableName));
