+++ schemacrawler/src/main/java/schemacrawler/tools/operation/OperatorHTMLOutput.java
@@ -110,7 +110,7 @@ final class OperatorHTMLOutput
    *      java.lang.String, java.lang.String, int, java.lang.String)
    */
   public void handleTable(final int ordinalPosition, final String tableName,
-                          final String tableType, final int count,
+                          final String tableType, final long count,
                           final String message)
   {
